Add minor commentary
Add minor commentary to clarify how the program is operating. The combat system is still very primative, and use health_points as the determining factor in combat. In the future, each object will get a power attribute that will determing it's strength in combat; and subsequently the amount of hitpoints lost by the opponent.